new features and reworked renderJournalSheet hook
- added a table of contents renderer
  - [toc] will be replaced with the rendered table of contents
  - entries are clickable and scroll the coresponding headline
    into view
- post table to chat now has a column filter
  - columns that have a header cell starting with a # get
    removed before posting to chat
