+++ module/module.json
@@ -1,45 +1,43 @@
 {
-  "name": "womboras-journal-helper",
-  "title": "Womboras Journal Helper",
-  "description": "A few helpers to make worldbuilding with journals a bit easier",
-  "author": "WooDWorkeR",
-  "authors": [
-    {
-      "name": "woodworker",
-      "discord": "WooDWorkeR#9185"
-    }
-  ],
-  "version": "0.0.4",
-  "minimumCoreVersion": "0.8.6",
-  "compatibleCoreVersion": "0.8.9",
-  "manifestPlusVersion": "1.0.0",
-  "url": "https://github.com/woodworker/womboras-journal-helper",
-  "manifest": "https://github.com/woodworker/womboras-journal-helper/releases/latest/download/module.json",
-  "download": "https://github.com/woodworker/womboras-journal-helper/releases/latest/download/womboras-journal-helper.zip",
-  "type": "module",
-  "media": [
-    {
-      "type": "icon",
-      "location": ""
-    }, 
-    {
-      "type": "cover",
-      "location": ""
-    },
-    {
-      "type": "screenshot",
-      "location": ""
-    }
-  ],
-  "esmodules": [
-    "module.js"
-  ],
-  "styles": [],
-  "languages": [
-    {
-      "lang": "en",
-      "name": "English",
-      "path": "lang/en.json"
-    }
-  ]
+    "name": "womboras-journal-helper",
+    "title": "Womboras Journal Helper",
+    "description": "A few helpers to make worldbuilding with journals a bit easier",
+    "author": "WooDWorkeR",
+    "authors": [
+        {
+            "name": "woodworker",
+            "discord": "WooDWorkeR#9185"
+        }
+    ],
+    "version": "0.0.5",
+    "minimumCoreVersion": "0.8.6",
+    "compatibleCoreVersion": "0.8.9",
+    "manifestPlusVersion": "1.0.0",
+    "url": "https://github.com/woodworker/womboras-journal-helper",
+    "manifest": "https://github.com/woodworker/womboras-journal-helper/releases/latest/download/module.json",
+    "download": "https://github.com/woodworker/womboras-journal-helper/releases/latest/download/womboras-journal-helper.zip",
+    "type": "module",
+    "media": [
+        {
+            "type": "icon",
+            "location": ""
+        },
+        {
+            "type": "cover",
+            "location": ""
+        },
+        {
+            "type": "screenshot",
+            "location": ""
+        }
+    ],
+    "esmodules": ["module.js"],
+    "styles": ["styles/module.css"],
+    "languages": [
+        {
+            "lang": "en",
+            "name": "English",
+            "path": "lang/en.json"
+        }
+    ]
 }
